+++ test/unit/core/VelodromeUnitTest.t.sol
@@ -10,6 +10,10 @@ import {uint512/*, uint512_external, alloc*/} from "src/utils/512Math.sol";
 import {Test} from "@forge-std/Test.sol";
 
 contract VelodromeConvergenceDummy is Velodrome {
+    function _tokenId() internal pure override returns (uint256) {
+        revert("unimplemented");
+    }
+
     function _msgSender() internal pure override returns (address) {
         revert("unimplemented");
     }