+++ CHANGELOG.md
@@ -39,6 +39,10 @@ Master list of UniV3 forks:
 
 ### Non-breaking changes
 
+* Create new Settler version for intent-based settlement -- the taker only signs
+  the slippage, not the actions
+  * This is now `tokenId` 4
+* Add UniswapV3 UniV3 fork to Monad Testnet
 * Add UniswapV4 actions to Sepolia
 * Add UniswapV4 actions to Ink